Fix for crash on 64-bit systems with ifort:  use 'size_t' instead of 'int' for the implicit string length parameter of Fortran functions.

+++ lithomop3d/module/lithomop3d_externs.h
@@ -31,6 +31,8 @@
 #if !defined(lithomop3d_externs_h)
 #define lithomop3d_externs_h
 
+#include <stddef.h>
+
 #define assign_wink_f FC_FUNC_(assign_wink, ASSIGN_WINK)
 #define autoprestr_f FC_FUNC(autoprestr, AUTOPRESTR)
 #define cmp_stiffsz_f FC_FUNC_(cmp_stiffsz, CMP_STIFFSZ)
@@ -203,10 +205,10 @@ extern "C" {
 		    int *,
 		    int *,             // error codes
 		    char *,
-		    int,               // string lengths
-		    int,
-		    int,
-		    int);
+		    size_t,               // string lengths
+		    size_t,
+		    size_t,
+		    size_t);
 
   void cmp_stiffsz_f(int *,
 	             int *,
@@ -216,7 +218,7 @@ extern "C" {
 	             int *,
 	             int *,
 	             int *,
-	             char *,int);
+	             char *,size_t);
 
   void create_id_f(int *,
 		   int *,
@@ -317,10 +319,10 @@ extern "C" {
                 int *,
 		int *,             // error codes
 		char *,
-		int,               // string lengths
-		int,
-		int,
-		int);
+		size_t,               // string lengths
+		size_t,
+		size_t,
+		size_t);
 
   void id_split_f(int *,
 		  int *,
@@ -343,7 +345,7 @@ extern "C" {
 		int *,
 		int *,
 		char *,
-		int);
+		size_t);
 
   void local_f(int *,
 	       int *,
@@ -401,7 +403,7 @@ extern "C" {
 		  int *,
 		  int *,
 		  int *,
-		  char *,int);
+		  char *,size_t);
 
   void read_bc_f(double*,
 		 double*,
@@ -415,8 +417,8 @@ extern "C" {
 		 char *,
 		 int *,
 		 char *,
-		 int,
-		 int);
+		 size_t,
+		 size_t);
 
   void read_connect_f(int *,
 		      int *,
@@ -428,8 +430,8 @@ extern "C" {
 		      char *,
 		      int *,
 		      char *,
-		      int,
-		      int);
+		      size_t,
+		      size_t);
 
   void read_coords_f(double*,
 		     double*,
@@ -438,8 +440,8 @@ extern "C" {
 		     char *,
 		     int *,
 		     char *,
-		     int,
-		     int);
+		     size_t,
+		     size_t);
 
   void read_diff_f(double*,
 		   int *,
@@ -451,8 +453,8 @@ extern "C" {
 		   char *,
 		   int *,
 		   char *,
-		   int,
-		   int);
+		   size_t,
+		   size_t);
 
   void read_fuldat_f(int *,
 		     int *,
@@ -463,8 +465,8 @@ extern "C" {
 		     char *,
 		     int *,
 		     char *,
-		     int,
-		     int);
+		     size_t,
+		     size_t);
 
   void read_hist_f(double *,
 		   double *,
@@ -474,8 +476,8 @@ extern "C" {
 		   char *,
 		   int *,
 		   char *,
-		   int,
-		   int);
+		   size_t,
+		   size_t);
 
   // void read_mathist_f(int *,
 // 		      int *,
@@ -492,10 +494,10 @@ extern "C" {
 // 		      char *,
 // 		      int *,
 // 		      char *,
-// 		      int,
-// 		      int,
-// 		      int,
-// 		      int);
+// 		      size_t,
+// 		      size_t,
+// 		      size_t,
+// 		      size_t);
 
   // void read_prestr_f(double *,
 		     // double *,
@@ -512,8 +514,8 @@ extern "C" {
 		     // int *,
 		     // char *,
 		     // char *,
-		     // int,
-		     // int);
+		     // size_t,
+		     // size_t);
 
   void read_skew_f(double *,
 		   double *,
@@ -524,8 +526,8 @@ extern "C" {
 		   char *,
 		   int *,
 		   char *,
-		   int,
-		   int);
+		   size_t,
+		   size_t);
 
   void read_slip_f(int *,
 		   int *,
@@ -536,8 +538,8 @@ extern "C" {
 		   char *,
 		   int *,
 		   char *,
-		   int,
-		   int);
+		   size_t,
+		   size_t);
 
   void read_split_f(double *,
 		    int *,
@@ -549,8 +551,8 @@ extern "C" {
 		    char *,
 		    int *,
 		    char *,
-		    int,
-		    int);
+		    size_t,
+		    size_t);
 
   void read_stateout_f(int *,
 		       int *,
@@ -558,8 +560,8 @@ extern "C" {
 		       char *,
 		       int *,
 		       char *,
-		       int,
-		       int);
+		       size_t,
+		       size_t);
 
   void read_timdat_f(double *,
 		     double *,
@@ -579,8 +581,8 @@ extern "C" {
 		     char *,
 		     int *,
 		     char *,
-		     int,
-		     int);
+		     size_t,
+		     size_t);
 
   // void read_traction_f(double *,
 		       // double *,
@@ -597,8 +599,8 @@ extern "C" {
 		       // int *,
 		       // char *,
 		       // char *,
-		       // int,
-		       // int);
+		       // size_t,
+		       // size_t);
 
   void read_wink_f(double *,
 		   double *,
@@ -610,8 +612,8 @@ extern "C" {
 		   char *,
 		   int *,
 		   char *,
-		   int,
-		   int);
+		   size_t,
+		   size_t);
 
 //   void read_winkx_f(double *,
 // 		    double *,
@@ -627,9 +629,9 @@ extern "C" {
 // 		    char *,
 // 		    int *,
 // 		    char *,
-// 		    int,
-// 		    int,
-// 		    int);
+// 		    size_t,
+// 		    size_t,
+// 		    size_t);
 
   void scan_bc_f(int *,
 		 int *,
@@ -639,11 +641,11 @@ extern "C" {
 		 char *,
 		 int *,
 		 char *,
-		 int,
-		 int,
-		 int,
-		 int,
-		 int);
+		 size_t,
+		 size_t,
+		 size_t,
+		 size_t,
+		 size_t);
 
   void scan_connect_f(int *,
 		      int *,
@@ -658,8 +660,8 @@ extern "C" {
 		      char *,
 		      int *,
 		      char *,
-		      int,
-		      int);
+		      size_t,
+		      size_t);
 
   void scan_coords_f(int *,
 		     int *,
@@ -667,9 +669,9 @@ extern "C" {
 		     char *,
 		     int *,
 		     char *,
-		     int,
-		     int,
-		     int);
+		     size_t,
+		     size_t,
+		     size_t);
 
   void scan_diff_f(int *,
 		   int *,
@@ -677,8 +679,8 @@ extern "C" {
 		   char *,
 		   int *,
 		   char *,
-		   int,
-		   int);
+		   size_t,
+		   size_t);
 
   void scan_fuldat_f(int *,
 		     int *,
@@ -687,16 +689,16 @@ extern "C" {
 		     char *,
 		     int *,
 		     char *,
-		     int,
-		     int);
+		     size_t,
+		     size_t);
 
   void scan_hist_f(int *,
 		   int *,
 		   char *,
 		   int *,
 		   char *,
-		   int,
-		   int);
+		   size_t,
+		   size_t);
 
   // void scan_prestr_f(int *,
 		     // int *,
@@ -705,7 +707,7 @@ extern "C" {
 		     // int *,
 		     // int *,
 		     // int *,
-		     // char *,int);
+		     // char *,size_t);
 
   void scan_skew_f(int *,
 		   int *,
@@ -713,25 +715,25 @@ extern "C" {
 		   char *,
 		   int *,
 		   char *,
-		   int,
-		   int,
-		   int);
+		   size_t,
+		   size_t,
+		   size_t);
 
   void scan_slip_f(int *,
 		   int *,
 		   char *,
 		   int *,
 		   char *,
-		   int,
-		   int);
+		   size_t,
+		   size_t);
 
   void scan_split_f(int *,
 		    int *,
 		    char *,
 		    int *,
 		    char *,
-		    int,
-		    int);
+		    size_t,
+		    size_t);
 
   void scan_timdat_f(int *,
 		     int *,
@@ -740,9 +742,9 @@ extern "C" {
 		     char *,
 		     int *,
 		     char *,
-		     int,
-		     int,
-		     int);
+		     size_t,
+		     size_t,
+		     size_t);
 
   // void scan_traction_f(int *,
 		       // int *,
@@ -751,8 +753,8 @@ extern "C" {
 		       // int *,
 		       // char *,
 		       // char *,
-		       // int,
-		       // int);
+		       // size_t,
+		       // size_t);
 
   void scan_wink_f(int *,
 		   int *,
@@ -760,8 +762,8 @@ extern "C" {
 		   char *,
 		   int *,
 		   char *,
-		   int,
-		   int);
+		   size_t,
+		   size_t);
 
   void scan_winkx_f(int *,
 		    int *,
@@ -770,8 +772,8 @@ extern "C" {
 		    char *,
 		    int *,
 		    char *,
-		    int,
-		    int);
+		    size_t,
+		    size_t);
 
   void sort_elements_f(int *,
 		       int *,
@@ -793,7 +795,7 @@ extern "C" {
 		       int *,
 		       int *,
 		       char *,
-		       int);
+		       size_t);
 
   void sort_slip_nodes_f(int *,
 		         int *,
@@ -805,7 +807,7 @@ extern "C" {
 		          int *,
 		          int *);
 
-  void try_binio_f(int *, int *, char *, int);
+  void try_binio_f(int *, int *, char *, size_t);
 
   void viscos_f(Mat *,             // sparse
 		double *,          // force
@@ -897,10 +899,10 @@ extern "C" {
                 int *,
 		int *,             // error codes
 		char *,
-		int,               // string lengths
-		int,
-		int,
-		int);
+		size_t,               // string lengths
+		size_t,
+		size_t,
+		size_t);
 
   void write_bc_f(double *,
 		  int *,
@@ -910,8 +912,8 @@ extern "C" {
 		  char *,
 		  int *,
 		  char *,
-		  int,
-		  int);
+		  size_t,
+		  size_t);
 
   void write_connect_f(int *,
 		       int *,
@@ -929,9 +931,9 @@ extern "C" {
 		       char *,
 		       int *,
 		       char *,
-		       int,
-		       int,
-		       int);
+		       size_t,
+		       size_t,
+		       size_t);
 
   void write_coords_f(double *,
 		      int *,
@@ -943,9 +945,9 @@ extern "C" {
 		      char *,
 		      int *,
 		      char *,
-		      int,
-		      int,
-		      int);
+		      size_t,
+		      size_t,
+		      size_t);
 
   void write_diff_f(double *,
 		    int *,
@@ -958,8 +960,8 @@ extern "C" {
 		    char *,
 		    int *,
 		    char *,
-		    int,
-		    int);
+		    size_t,
+		    size_t);
 
   void write_element_info_f(int *,
 			    int *,
@@ -972,7 +974,7 @@ extern "C" {
 			    double *,
 			    int *,
 			    int *,
-			    char *,int);
+			    char *,size_t);
 
   void write_fuldat_f(int *,
 		      int *,
@@ -987,9 +989,9 @@ extern "C" {
 		      char *,
 		      int *,
 		      char *,
-		      int,
-		      int,
-		      int);
+		      size_t,
+		      size_t,
+		      size_t);
 
   void write_global_info_f(char *,
 			   int *,
@@ -1001,9 +1003,9 @@ extern "C" {
 			   int *,
 			   char *,
 			   char *,
-			   int,
-			   int,
-			   int);
+			   size_t,
+			   size_t,
+			   size_t);
 
   void write_hist_f(double *,
 		    double *,
@@ -1014,8 +1016,8 @@ extern "C" {
 		    char *,
 		    int *,
 		    char *,
-		    int,
-		    int);
+		    size_t,
+		    size_t);
 
   void write_props_f(double *,
 		     double *,
@@ -1031,9 +1033,9 @@ extern "C" {
 		     char *,
 		     int *,
 		     char *,
-		     int,
-		     int,
-		     int);
+		     size_t,
+		     size_t,
+		     size_t);
 
   void write_skew_f(double *,
 		    int *,
@@ -1044,8 +1046,8 @@ extern "C" {
 		    char *,
 		    int *,
 		    char *,
-		    int,
-		    int);
+		    size_t,
+		    size_t);
 
   void write_slip_f(int *,
 		    int *,
@@ -1058,9 +1060,9 @@ extern "C" {
 		    char *,
 		    int *,
 		    char *,
-		    int,
-		    int,
-		    int);
+		    size_t,
+		    size_t,
+		    size_t);
 
   void write_sparse_info_f(int *,
 			   int *,
@@ -1069,7 +1071,7 @@ extern "C" {
 			   double *,
 			   int *,
 			   int *,
-			   char *,int);
+			   char *,size_t);
 
   void write_split_f(double *,
 		     int *,
@@ -1082,15 +1084,15 @@ extern "C" {
 		     char *,
 		     int *,
 		     char *,
-		     int,
-		     int,
-		     int);
+		     size_t,
+		     size_t,
+		     size_t);
 
   void write_split_plot_f(int *,
 		          int *,
 		          int *,
 		          int *,
-			  char *,int);
+			  char *,size_t);
 
   void write_stateout_f(int *,
 		        int *,
@@ -1102,21 +1104,21 @@ extern "C" {
 			char *,
 			int *,
 			char *,
-			int,
-			int,
-			int);
+			size_t,
+			size_t,
+			size_t);
 
   void write_strscomp_f(double *,
 			double *,
 			double *,
 			int *,
 			int *,
-			char *,int);
+			char *,size_t);
 
   void write_subiter_f(int *,
 		       int *,
 		       int *,
-		       char *,int);
+		       char *,size_t);
 
   void write_timdat_f(double *,
 		      double *,
@@ -1136,8 +1138,8 @@ extern "C" {
 		      char *,
 		      int *,
 		      char *,
-		      int,
-		      int);
+		      size_t,
+		      size_t);
 
   void write_ucd_mesh_f(double *,
 		       	int *,
@@ -1153,7 +1155,7 @@ extern "C" {
 		       	int *,
 		       	int *,
 		       	int *,
-		       	char *,int);
+		       	char *,size_t);
 
   void write_wink_f(double *,
 		    int *,
@@ -1164,8 +1166,8 @@ extern "C" {
 		    char *,
 		    int *,
 		    char *,
-		    int,
-		    int);
+		    size_t,
+		    size_t);
 
   void write_winkx_f(double *,
 		     int *,
@@ -1176,8 +1178,8 @@ extern "C" {
 		     char *,
 		     int *,
 		     char *,
-		     int,
-		     int);
+		     size_t,
+		     size_t);
 }
 
 #endif // lithomop3d_externs_h
